Update README.md
Why do you use "lbl" as a key for the term/class name? Why not "name"? Also, I would be in favor of using 
"subject", "predicate" and "object" instead of the three-character abbreviations. It'll be just a tad clearer.

I would add in the motivation-section that JSON is a very popular syntax that has a lot of support in various software. 
It is more expressive than the OBO syntax and less complicated than OWL.

OBO is a subset of OWL, i.e. OWL is a superset thus can express all axioms found in OBO. How does OBO Graph compare with OWL? can they both express each other's axioms and properties? Are they equivalent? Can I convert from one to the other and back and arrive at the same state?